Aplikasi/requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
astor==0.8.1
async-timeout==4.0.3
attrs==25.3.0
base58==2.1.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
cachetools==6.0.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
cftime==1.6.4.post1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.0
cryptography==45.0.4
cycler==0.12.1
cymem==2.0.11
dan==0.0.0
datasets==3.6.0
dill==0.3.8
distro==1.9.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.2
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kbbi==0.4.3
kiwisolver==1.4.7
langchain==0.3.26
langchain-core==0.3.72
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.8
language_data==1.3.0
llvmlite==0.43.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==1.40.0
netCDF4==1.7.2
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
ollama==0.5.1
openai==1.82.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.11.0
outcome==1.3.0.post0
packaging==25.0
pandas==1.5.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
preshed==3.0.10
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
pynndescent==0.5.13
pyparsing==3.2.3
pypdfium2==4.30.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.33.0
semver==3.0.4
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
streamlit==1.11.0
streamlit-ace==0.1.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
urllib3==2.4.0
validators==0.35.0
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xarray==2024.7.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
